fix(alerts): cast explícito de bp_multiplier/bp_min_days en broken_pattern
PG inferia $1 (bp_multiplier) como integer porque GREATEST con
COALESCE($2, 0) lo unificaba al tipo int del literal 0. bp_multiplier viene
del DB como NUMERIC y node-pg lo serializa como string "2.00" — que no
parsea como integer y rompe el runner en producción:

  ERROR: invalid input syntax for type integer: "2.00"

Fix: castear explícitamente cs.avg_interval_days * $1::numeric y
COALESCE($2::int, 0), y redondear el producto a int para que el GREATEST
sea entre dos ints comparables a lv.days_since_last (int).

+++ workshop-backend/src/services/alertStrategies.js
@@ -251,7 +251,7 @@ const brokenPattern = {
       JOIN last_visit lv ON lv.client_id = cs.client_id
       JOIN clients c     ON c.id = cs.client_id
       WHERE c.deleted_at IS NULL
-        AND lv.days_since_last > GREATEST(cs.avg_interval_days * $1, COALESCE($2, 0))
+        AND lv.days_since_last > GREATEST((cs.avg_interval_days * $1::numeric)::int, COALESCE($2::int, 0))
       ORDER BY (lv.days_since_last::float / cs.avg_interval_days) DESC
       LIMIT 100
     `, [def.bp_multiplier, def.bp_min_days]);
